python/requirements.txt
bit >= 0.6.3
bitcoinlib >= 0.4.16
python_bitcoinlib >= 0.11.0
numpy >= 1.15.4
bitpost >= 1.0.4